Everestinit

+++ Models/AddInOrderDetail.cs
@@ -5,11 +5,11 @@ namespace Sips.Models;
 
 public partial class AddInOrderDetail
 {
-    public long FkAddInId { get; set; }
+    public int FkAddInId { get; set; }
 
-    public long FkOrderDetailId { get; set; }
+    public int FkOrderDetailId { get; set; }
 
-    public long Quantity { get; set; }
+    public int Quantity { get; set; }
 
     public virtual AddIn FkAddIn { get; set; } = null!;
